requirements.txt
Brotli==1.0.9
click==8.0.1
configparser==5.0.2
dash==1.21.0
dash-bootstrap-components==0.13.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
Flask==2.0.1
Flask-Compress==1.10.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.1
pandas==1.3.1
patsy==0.5.1
plotly==5.1.0
plotly-express==0.4.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.22
statsmodels==0.10.2
tenacity==8.0.1
Werkzeug==2.0.1
XlsxWriter==3.0.1
